% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{CS}
\alias{CS}
\alias{AR}
\title{Design matrix generators}
\usage{
CS(n, p, rho)

AR(n, p, rho)
}
\arguments{
\item{n}{number of rows}

\item{p}{number of columns}

\item{rho}{correlation parameter}
}
\description{
Design matrix generators
}
