% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{est_lambda}
\alias{est_lambda}
\title{Multiplier-bootstrap based lambda estimator (rank-style)}
\usage{
est_lambda(X, alpha0 = 0.1, c = 1.01, times = 100)
}
\arguments{
\item{X}{design matrix}

\item{alpha0}{1 - confidence level}

\item{c}{inflation factor}

\item{times}{number of draws}
}
\description{
Multiplier-bootstrap based lambda estimator (rank-style)
}
