% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matsplitter}
\alias{matsplitter}
\title{Matrix splitter (internal utility)}
\usage{
matsplitter(M, r, c)
}
\arguments{
\item{M}{matrix to split}

\item{r}{number of row blocks}

\item{c}{number of column blocks}
}
\value{
A nested list indexed as \link{row_block, col_block}
}
\description{
Matrix splitter (internal utility)
}
