% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prox_beta}
\alias{prox_beta}
\alias{prox_z}
\title{Proximal and utility helpers}
\usage{
prox_beta(u, a)

prox_z(xi, alpha, quan)
}
\arguments{
\item{u, a}{numeric vectors}

\item{xi, alpha}{numeric vectors}

\item{quan}{quantile level in (0,1)}
}
\description{
Proximal and utility helpers
}
